refactor(common): move `EnumUtil` to `util` package in common module
